Test X is not copied, and test deprecation warning.
