+++ anndata/tests/test_concatenate.py
@@ -17,16 +17,8 @@ from anndata import AnnData, Raw, concat
 from anndata._core.index import _subset
 from anndata._core import merge
 from anndata.tests import helpers
-from anndata.tests.helpers import assert_equal, gen_adata
+from anndata.tests.helpers import assert_equal, gen_adata, array_type
 from anndata.utils import asarray
-
-
-@pytest.fixture(
-    params=[asarray, sparse.csr_matrix, sparse.csc_matrix],
-    ids=["np_array", "scipy_csr", "scipy_csc"],
-)
-def array_type(request):
-    return request.param
 
 
 @pytest.fixture(params=["inner", "outer"])
@@ -863,10 +855,7 @@ def gen_concat_params(unss, compat2result):
             },
         ),
         gen_concat_params(
-            [
-                {"a": {"b": 1, "c": {"d": 3}}},
-                {"a": {"b": 1, "c": {"e": 4}}},
-            ],
+            [{"a": {"b": 1, "c": {"d": 3}}}, {"a": {"b": 1, "c": {"e": 4}}},],
             {
                 None: {},
                 "first": {"a": {"b": 1, "c": {"d": 3, "e": 4}}},
@@ -876,12 +865,7 @@ def gen_concat_params(unss, compat2result):
             },
         ),
         gen_concat_params(
-            [
-                {"a": 1},
-                {"a": 1, "b": 2},
-                {"a": 1, "b": {"b.a": 1}, "c": 3},
-                {"d": 4},
-            ],
+            [{"a": 1}, {"a": 1, "b": 2}, {"a": 1, "b": {"b.a": 1}, "c": 3}, {"d": 4},],
             {
                 None: {},
                 "first": {"a": 1, "b": 2, "c": 3, "d": 4},
